Lint accordingly
